Raf session

+++ wtw-frontend/src/components/Users.jsx
@@ -13,7 +13,8 @@ class Users extends Component {
             patient: context.state.patient,
             patients: [],
             nom: '',
-            prenom: ''
+            prenom: '',
+            sessions: []
         }
         this.updateSession = context.updateSession.bind(this);
         this.isActive = this.isActive.bind(this);
@@ -26,7 +27,19 @@ class Users extends Component {
 
     session = user => {
         this.updateSession(user)
-        this.setState({ patient: user })
+        this.setState({ 
+            patient: user,
+            sessions: []
+                      })
+    
+        
+        api.getSession(user._id).then(resp => {
+            resp.data.map(session => this.setState(prevState => ({
+                sessions: [...prevState.sessions, session]
+            }))
+                         )
+        })
+
     }
 
     infoPatient = () => {
@@ -38,10 +51,10 @@ class Users extends Component {
                 <Row className="mt-4 w-100">
                     <Button variant="danger" className="ml-auto"
                         onClick={() =>
-                            window.confirm("Voulez vous vraiment supprimer ce patient?") &&
-                            this.handleDeleteUser()
-                        }
-                    >
+                window.confirm("Voulez vous vraiment supprimer ce patient?") &&
+                    this.handleDeleteUser()
+                                }
+                        >
                         Supprimer patient
                     </Button>
                 </Row>
@@ -49,42 +62,124 @@ class Users extends Component {
         )
     }
 
-    sessionPatient = () => {
-        const options = {
-            chart: {
-                zoom: {
-                    enabled: false
-                }
-            },
-            xaxis: {
-                categories: ['Jan', 'Feb', 'Mar', 'Apr', 'Mai']
-            },
-            title: {
-                text: 'Sessions Louis Marquis',
-                align: 'left'
-            },
-        };
+    //    sessionPatient = async() => {
+    //
+    //        await api.getSession(this.state.patient._id).then(sessions => { 
+    //
+    //            const items = [];
+    //            var options = null;
+    //            var series = null;
+    //
+    //            for (const [index, value] of sessions.data.entries()) {
+    //                options = {
+    //                    chart: {
+    //                        zoom: {
+    //                            enabled: false
+    //                        }
+    //                    },
+    //                    xaxis: {
+    //                        categories: [value.debut, value.fin]
+    //                    },
+    //                    title: {
+    //                        text: 'Sessions '+this.state.patient.nom+' '+this.state.patient.prenom,
+    //                        align: 'left'
+    //                    },
+    //                };
+    //
+    //                series = [{
+    //                    name: "Angle",
+    //                    data: value.data
+    //                },
+    //                          {
+    //                              name: "Poids",
+    //                              data: value.data
+    //                          }]
+    //                items.push(
+    //                    <LineChart key={index} series={series} options={options} />
+    //                )
+    //            }
+    //
+    //            console.log(items)
+    //
+    //            return (<Row className="w-100">
+    //                    {items}
+    //                </Row>)
+    //        })
+    //
+    //    }
 
-        const series = [{
-            name: "Angle",
-            data: [50, 60, 70, 80, 85]
-        }]
+//    sessionPatient = () => {
+//        api.getSession(this.state.patient._id).then(sessions=>{
+//            console.log(sessions.data)
+//            return (
+//                <Row className="w-100">
+//                    {sessions.data.map(session => {
+//                        var options = {
+//                            chart: {
+//                                zoom: {
+//                                    enabled: false
+//                                }
+//                            },
+//                            xaxis: {
+//                                categories: [session.debut, session.fin]
+//                            },
+//                            title: {
+//                                text: 'Sessions '+this.state.patient.nom+' '+this.state.patient.prenom,
+//                                align: 'left'
+//                            },
+//                        };
+//
+//                        var series = [{
+//                            name: "Angle",
+//                            data: session.data
+//                        },
+//                                      {
+//                                          name: "Poids",
+//                                          data: session.data
+//                                      }]
+//                        return <LineChart key={series} options={options} />
+//                    })}
+//                </Row>
+//            )
+//        })    
+//    }
 
-        return (
-            <Row className="w-100">
-                {/* TODO: map sessions */}
-                <LineChart series={series} options={options} />
-            </Row>
-        )
-    }
+            sessionPatient = () => {
+                
+//                const options = {
+//                    chart: {
+//                        zoom: {
+//                            enabled: false
+//                        }
+//                    },
+//                    xaxis: {
+//                        categories: ['Jan', 'Feb', 'Mar', 'Apr', 'Mai']
+//                    },
+//                    title: {
+//                        text: 'Sessions Louis Marquis',
+//                        align: 'left'
+//                    },
+//                };
+//        
+//                const series = [{
+//                    name: "Angle",
+//                    data: [50, 60, 70, 80, 85]
+//                }]
+//        
+//                return (
+//                    <Row className="w-100">
+//                        {/* TODO: map sessions */}
+//                        <LineChart series={series} options={options} />
+//                    </Row>
+//                )
+              }
 
     componentDidMount = async () => {
         await api.getAllUsers().then(resp => {
-            resp.data.map(user =>
-                this.setState(prevState => ({
-                    patients: [...prevState.patients, user]
-                }))
-            )
+            resp.data.map(user => this.setState(prevState => ({
+                patients: [...prevState.patients, user]
+            }))
+                         )
         })
     }
 
@@ -116,15 +211,15 @@ class Users extends Component {
                     <Button
                         className={this.isActive({ _id: '' })} onClick={() => this.session({ _id: '' })}
                         variant="outline-secondary"
-                    >
+                        >
                         Sans patient
                     </Button>
 
                     {this.state.patients.map(patient => {
                         return <Button
-                            className={this.isActive(patient)} onClick={() => this.session(patient)}
-                            key={patient._id} variant="outline-primary"
-                        > {patient.nom} {patient.prenom}
+                                   className={this.isActive(patient)} onClick={() => this.session(patient)}
+                                   key={patient._id} variant="outline-primary"
+                                   > {patient.nom} {patient.prenom}
                         </Button>
                     })}
 
@@ -159,6 +254,39 @@ class Users extends Component {
                 </Accordion>
 
                 {this.state.patient.nom ? this.infoPatient() : null}
+                
+                {this.state.sessions.map(session=>{
+                    var options = {
+                    chart: {
+                        zoom: {
+                            enabled: false
+                        }
+                    },
+                    xaxis: {
+                        categories: [session.debut,session.fin]
+                    },
+                    title: {
+                        text: 'Sessions '+this.state.patient.nom+' '+this.state.patient.prenom,
+                        align: 'left'
+                    },
+                };
+        
+                var series = [{
+                    name: "Angle",
+                    data: session.data
+                },{
+                    name: "Poids",
+                    data: session.data
+                }]
+        
+                return (
+                    <Row className="w-100">
+                        {/* TODO: map sessions */}
+                        <LineChart series={series} options={options} />
+                    </Row>
+                )
+                    
+                })}
 
             </Container>
         );
